[Patch v2] Return ChildProcess to caller
The utility of get-exec-file is somewhat limited if the created
ChildProcess is not returned to the caller, which can no longer get the
child PID, write to or close stdin, or send signals.  To avoid this
limitation without breaking the current API, return the ChildProcess as
a property on the returned Promise.

Changes in v2:
- Return ChildProcess with then and catch from Promise, as requested.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>